# docs/swagger.yaml
definitions:
  wallet.Err:
    properties:
      message:
        type: string
    type: object
  wallet.Wallet:
    properties:
      balance:
        example: 100
        type: number
      created_at:
        example: "2024-03-25T14:19:00.729237Z"
        type: string
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        type: string
      wallet_name:
        example: John's Wallet
        type: string
      wallet_type:
        example: Credit Card
        type: string
    type: object
  wallet.WalletForCreate:
    properties:
      balance:
        example: 100
        type: number
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        type: string
      wallet_name:
        example: John's Wallet
        type: string
      wallet_type:
        example: Credit Card
        type: string
    type: object
  wallet.WalletForUpdate:
    properties:
      balance:
        example: 100
        type: number
      id:
        example: 1
        type: integer
    type: object
host: localhost:1323
info:
  contact: {}
  description: Sophisticated Wallet API
  title: Wallet API
  version: "1.0"
paths:
  /api/v1/user/{id}/wallets:
    delete:
      description: Delete wallet for the user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wallet.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Delete wallet for the user
      tags:
      - user wallet
  /api/v1/users/{id}/wallets:
    get:
      consumes:
      - application/json
      description: Get all wallets for the user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/wallet.Wallet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wallet.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get all wallets for the user
      tags:
      - user wallet
  /api/v1/wallets:
    get:
      consumes:
      - application/json
      description: Get all wallets
      parameters:
      - description: Filter by wallet type
        in: query
        name: wallet_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/wallet.Wallet'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get all wallets
      tags:
      - wallet
    post:
      consumes:
      - application/json
      description: Create wallet
      parameters:
      - description: Wallet object
        in: body
        name: wallet
        required: true
        schema:
          $ref: '#/definitions/wallet.WalletForCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wallet.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Create wallet
      tags:
      - wallet
    put:
      consumes:
      - application/json
      description: Update wallet
      parameters:
      - description: Wallet object
        in: body
        name: wallet
        required: true
        schema:
          $ref: '#/definitions/wallet.WalletForUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wallet.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Update wallet
      tags:
      - wallet
swagger: "2.0"